ui: add icons + refactor some code

+++ app/Livewire/NoteDestroy.php
@@ -3,6 +3,7 @@
 namespace App\Livewire;
 
 use App\Models\Note;
+use Illuminate\Support\Facades\Log;
 use Livewire\Component;
 
 class NoteDestroy extends Component
@@ -15,8 +16,19 @@ class NoteDestroy extends Component
     public function destroy($slug)
     {
         $note = Note::whereSlug($slug)->first();
-        $note->delete();
 
-        return $this->redirect('/note', navigate: true);
+        if ($note->user_id === auth()->user()->id) {
+            try {
+                $note->delete();
+
+                return $this->redirect('/note', navigate: true);
+            } catch (\Throwable $err) {
+                Log::error($err->getMessage());
+
+                $this->dispatch('nyatetError');
+            }
+        } else {
+            $this->dispatch('nyatetNotMine');
+        }
     }
 }
